[YP-32] move types
