+++ packages/client/src/Store/Slices/Game/Game.slice.ts
@@ -1,49 +1,6 @@
 import type { PayloadAction } from '@reduxjs/toolkit';
 import { createSlice } from '@reduxjs/toolkit';
-
-export type TGameState = 'start' | 'game' | 'end';
-export enum EnemyType {
-  asteroid = 'asteroid',
-  bomb = 'bomb',
-  pirate = 'pirate',
-}
-
-export type TPlayer = {
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-  speed: number;
-};
-
-export type TBaseEnemy = {
-  id: string;
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-  speed: number;
-  type: EnemyType;
-};
-
-export type TPirate = TBaseEnemy & {
-  coefficient: 1 | -1;
-};
-
-export type TBomb = TBaseEnemy & {
-  coefficient: number;
-};
-
-export type TEnemy = TBaseEnemy | TPirate | TBomb;
-
-export type IGameInitState = {
-  gameState: TGameState;
-  gameSpeed: number;
-  player: TPlayer;
-  enemies: TEnemy[];
-  maxEnemyCount: number;
-  hitsCount: number;
-};
+import { IGameInitState, TEnemy, TGameState } from '@Store';
 
 const initialState: IGameInitState = {
   gameState: 'start',
@@ -56,7 +13,7 @@ const initialState: IGameInitState = {
     height: 54,
     speed: 1,
   },
-  maxEnemyCount: 3,
+  maxEnemyCount: 7,
   enemies: [],
 };
